# api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: Intelligent Transcription API
  description: |
    API REST pour le service de transcription audio/vidéo intelligente.
    
    ## Authentification
    L'API utilise des tokens JWT. Obtenez un token via `/auth/login` et incluez-le dans l'header:
    ```
    Authorization: Bearer YOUR_TOKEN
    ```
    
    ## Rate Limiting
    - 60 requêtes par minute
    - 1000 requêtes par heure
    
    ## Formats supportés
    - Audio: MP3, WAV, MP4, M4A
    - Vidéo: MP4, YouTube URLs
    - Langues: fr, en, es, de, it, pt, nl, pl, ru, ja, zh, ar
    
  version: 2.0.0
  contact:
    email: api@transcription.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v2
    description: Production API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transcription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        language:
          type: object
          properties:
            code:
              type: string
              example: fr
            name:
              type: string
              example: Français
        text:
          type: string
          description: Texte transcrit (disponible si status = completed)
        cost:
          type: object
          properties:
            amount:
              type: number
              format: float
            currency:
              type: string
              example: USD
        youtube:
          type: object
          properties:
            title:
              type: string
            video_id:
              type: string
            duration:
              type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /transcriptions:
    get:
      tags:
        - transcriptions
      summary: Lister les transcriptions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Liste des transcriptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transcription'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      per_page:
                        type: integer
                      current_page:
                        type: integer
                      total_pages:
                        type: integer
                        
    post:
      tags:
        - transcriptions
      summary: Créer une transcription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - language
              properties:
                audio:
                  type: string
                  format: binary
                  description: Fichier audio (requis si youtube_url absent)
                youtube_url:
                  type: string
                  description: URL YouTube (requis si audio absent)
                language:
                  type: string
                  description: Code de langue (fr, en, es, etc.)
      responses:
        '201':
          description: Transcription créée
          headers:
            Location:
              description: URL de la nouvelle ressource
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
...